Add Vagrant support
With this commit I add support for vagrant to setup a default development
environment. This installs a Ubuntu Precise Pangolin (64bit) box with
the necessary modules and applications to immediately start development on
phpDocumentor.

Please note that the first time you do 'vagrant up' that it will take a long
time to boot as composer needs to retrieve all dependencies.