Started specification for adding routers to phpDocumentor
Templates are capable of determining their own URL scheme. Although this is a
powerful feature it also means that behaviours do not know up front what the
URL scheme is going to look like.

To compensate for this it should become possible to define a mapping between
a FQSEN and a URL. This may be a class or defined inline with the template.

This commit contains the first series of specification changes