#636: Added EngineManager and refactored code
In this commit I cleaned up the code by adding DocBlocks and renaming
the Engine namespace to Adapter in order to be more explicit in our
intention with that code.

In order to provide a clear entry point and extension point I added an
EngineManager class, inspired on Doctrine's EntityManager, that calls
the correct adapter and works as a frontend.