Add JMS Serializer files to file-bin
Kevin Herrera of the Box project mentioned that the most likely
fix for this issue was by putting the JMS Serializer-based files
in a files-bin array int he box.json (https://github.com/box-project/box2/issues/85#issuecomment-57182168).

Based on his recommendation this is something that we are going
to try out.